ldexp requires i64 (not u64) exponent in numpy 1.21

+++ klamath/basic.py
@@ -43,7 +43,7 @@ def decode_real8(nums: numpy.ndarray) -> numpy.ndarray:
     exp = (nums >> 56) & 0x7f
     mant = (nums & 0x00ff_ffff_ffff_ffff).astype(numpy.float64)
     mant[neg != 0] *= -1
-    return numpy.ldexp(mant, 4 * (exp - 64) - 56, dtype=numpy.float64)
+    return numpy.ldexp(mant, (4 * (exp - 64) - 56).astype(numpy.int64))
 
 
 def parse_real8(data: bytes) -> numpy.ndarray:
